fix test and formatting
